Refactor user model and routes
